moved ResourceHandler to misc
