+++ app/src/main/java/org/techconnect/activities/GuideActivity.java
@@ -20,8 +20,8 @@ import android.widget.TextView;
 import com.squareup.picasso.Picasso;
 
 import org.centum.techconnect.R;
+import org.techconnect.misc.ResourceHandler;
 import org.techconnect.model.FlowChart;
-import org.techconnect.resources.ResourceHandler;
 import org.techconnect.services.TCService;
 import org.techconnect.sql.TCDatabaseHelper;
 import org.techconnect.views.CommentsResourcesTabbedView;
